Clean up tests for wrapup
