+++ tests/atlasgen/test_wrapup.py
@@ -1,8 +1,17 @@
-"""Unit tests for pure helper functions in wrapup.py."""
+"""
+Unit tests for pure helper functions in wrapup.py.
+
+Tests for overwrite and early-exit behaviour in wrapup_atlas_from_data.
+
+These tests verify that atlas generation fails early when output already
+exists, and that the overwrite flag correctly replaces existing output.
+"""
 
 import json
 
+import numpy as np
 import pandas as pd
+import pytest
 
 from brainglobe_atlasapi.atlas_generation.wrapup import (
     _build_transformations,
@@ -10,7 +19,9 @@ from brainglobe_atlasapi.atlas_generation.wrapup import (
     _save_coordinate_space_manifest,
     _save_if_not_exists,
     _save_terminology_csv,
+    wrapup_atlas_from_data,
 )
+from brainglobe_atlasapi.descriptors import ATLAS_VERSION
 
 # --- _merge_resolutions_list ---
 
@@ -64,7 +75,9 @@ def test_build_transformations_anisotropic():
 # --- _save_terminology_csv ---
 
 
-def _simple_structures():
+@pytest.fixture
+def simple_structures():
+    """Provide a simple list of structures for testing."""
     return [
         {
             "id": 999,
@@ -83,17 +96,17 @@ def _simple_structures():
     ]
 
 
-def test_save_terminology_csv_creates_file(tmp_path):
+def test_save_terminology_csv_creates_file(tmp_path, simple_structures):
     """Test that _save_terminology_csv creates the output CSV file."""
     csv_path = tmp_path / "terminology.csv"
-    _save_terminology_csv(_simple_structures(), csv_path)
+    _save_terminology_csv(simple_structures, csv_path)
     assert csv_path.exists()
 
 
-def test_save_terminology_csv_columns(tmp_path):
+def test_save_terminology_csv_columns(tmp_path, simple_structures):
     """Test that the CSV has the expected column names in the correct order."""
     csv_path = tmp_path / "terminology.csv"
-    _save_terminology_csv(_simple_structures(), csv_path)
+    _save_terminology_csv(simple_structures, csv_path)
     df = pd.read_csv(csv_path)
     assert list(df.columns) == [
         "identifier",
@@ -106,25 +119,27 @@ def test_save_terminology_csv_columns(tmp_path):
     ]
 
 
-def test_save_terminology_csv_root_has_no_parent(tmp_path):
+def test_save_terminology_csv_root_has_no_parent(tmp_path, simple_structures):
     """Test that the root structure has a NaN parent_identifier."""
     csv_path = tmp_path / "terminology.csv"
-    _save_terminology_csv(_simple_structures(), csv_path)
+    _save_terminology_csv(simple_structures, csv_path)
     df = pd.read_csv(csv_path)
     root_row = df[df["identifier"] == 999].iloc[0]
     assert pd.isna(root_row["parent_identifier"])
 
 
-def test_save_terminology_csv_child_has_correct_parent(tmp_path):
+def test_save_terminology_csv_child_has_correct_parent(
+    tmp_path, simple_structures
+):
     """Test that a child structure has the correct parent_identifier."""
     csv_path = tmp_path / "terminology.csv"
-    _save_terminology_csv(_simple_structures(), csv_path)
+    _save_terminology_csv(simple_structures, csv_path)
     df = pd.read_csv(csv_path)
     child_row = df[df["identifier"] == 1].iloc[0]
     assert child_row["parent_identifier"] == 999
 
 
-def test_save_terminology_csv_color_hex_format(tmp_path):
+def test_save_terminology_csv_color_hex_format(tmp_path, simple_structures):
     """Test that RGB triplets are correctly formatted as hex color strings."""
     structures = [
         {
@@ -203,3 +218,100 @@ def test_save_if_not_exists_skips_when_dir_exists(tmp_path):
         save_fn=fake_save,
     )
     assert calls == []
+
+
+@pytest.fixture
+def minimal_valid_inputs(tmp_path):
+    """
+    Return a minimal set of valid inputs required to run wrapup
+    past the overwrite logic without performing actual atlas gen.
+    """
+    return dict(
+        atlas_name="test_mouse",
+        atlas_minor_version=0,
+        citation="unpublished",
+        atlas_link="http://example.com",
+        species="Mouse (Mus musculus)",
+        resolution=(25, 25, 25),
+        orientation="asr",
+        root_id=0,
+        reference_stack=np.zeros((2, 2, 2)),
+        annotation_stack=np.zeros((2, 2, 2), dtype=np.uint32),
+        structures_list=[
+            {
+                "id": 0,
+                "acronym": "root",
+                "name": "root",
+                "rgb_triplet": [255, 255, 255],
+                "structure_id_path": [0],
+            }
+        ],
+        meshes_dict={},
+        working_dir=tmp_path,
+    )
+
+
+@pytest.fixture
+def expected_atlas_dir(minimal_valid_inputs):
+    """Return the path where wrapup writes the atlas manifest."""
+    atlas_version = f"{ATLAS_VERSION}.0".replace(".", "_")
+    return (
+        minimal_valid_inputs["working_dir"]
+        / "brainglobe-atlasapi"
+        / "atlases"
+        / "test_mouse_25um"
+        / atlas_version
+    )
+
+
+def test_wrapup_fails_if_output_exists(
+    expected_atlas_dir, minimal_valid_inputs
+):
+    """Fail early if atlas output already exists and overwrite=False."""
+    atlas_dir = expected_atlas_dir
+    atlas_dir.mkdir(parents=True)
+
+    kwargs = minimal_valid_inputs
+
+    with pytest.raises(FileExistsError, match="Atlas output already exists"):
+        wrapup_atlas_from_data(**kwargs, overwrite=False)
+
+
+def test_wrapup_overwrites_existing_output(
+    monkeypatch, expected_atlas_dir, minimal_valid_inputs
+):
+    """Overwrite existing atlas output when overwrite=True."""
+    atlas_dir = expected_atlas_dir
+    atlas_dir.mkdir(parents=True)
+    (atlas_dir / "old_file.txt").write_text("old")
+
+    from brainglobe_atlasapi.atlas_generation import wrapup
+
+    # Stub out heavy I/O that runs after the overwrite check.
+    class _FakeImage:
+        data = type("arr", (), {"shape": (2, 2, 2)})()
+
+    class _FakeMultiscale:
+        images = [_FakeImage()]
+
+    monkeypatch.setattr(
+        wrapup, "_save_template_data", lambda *a, **kw: _FakeMultiscale()
+    )
+    monkeypatch.setattr(
+        wrapup, "_save_annotation_data", lambda *a, **kw: (None, None)
+    )
+    monkeypatch.setattr(
+        wrapup, "_save_additional_references", lambda *a, **kw: None
+    )
+    monkeypatch.setattr(wrapup, "get_all_validation_functions", lambda: [])
+    monkeypatch.setattr(
+        wrapup,
+        "BrainGlobeAtlas",
+        lambda *args, **kwargs: None,
+    )
+
+    kwargs = minimal_valid_inputs
+    wrapup_atlas_from_data(**kwargs, overwrite=True)
+
+    assert atlas_dir.exists()
+    assert not (atlas_dir / "old_file.txt").exists()
